+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/MatrixMessengerConfiguration.kt
@@ -1,7 +1,6 @@
 package de.connect2x.trixnity.messenger
 
 import de.connect2x.trixnity.messenger.util.gb
-import de.connect2x.trixnity.messenger.util.kb
 import de.connect2x.trixnity.messenger.util.mb
 import org.koin.core.module.Module
 
@@ -50,7 +49,7 @@ data class MatrixMessengerConfiguration(
     /**
      * The maximum size of clipboard files that can be uploaded
      */
-    val clipboardMaxSize: Long = 1.gb(),
+    val filePreviewMaxSize: Long = 50.mb(),
 
     /**
      * The maximum size of avatars that can be uploaded in *Bytes*