+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/files/MediaViewModel.kt
@@ -1,12 +1,16 @@
+import de.connect2x.trixnity.messenger.MatrixMessengerConfiguration
 import de.connect2x.trixnity.messenger.i18n.I18n
 import de.connect2x.trixnity.messenger.util.FileTransferProgressElement
 import de.connect2x.trixnity.messenger.viewmodel.MatrixClientViewModelContext
 import de.connect2x.trixnity.messenger.viewmodel.room.timeline.OpenModalType
+import de.connect2x.trixnity.messenger.viewmodel.util.MaxByteFlowSizeException
 import de.connect2x.trixnity.messenger.viewmodel.util.formatProgress
+import de.connect2x.trixnity.messenger.viewmodel.util.limitSize
 import io.github.oshai.kotlinlogging.KotlinLogging
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
 import net.folivo.trixnity.client.media
@@ -71,6 +75,7 @@ open class MediaViewModelImpl(
         coroutineScope.launch {
             val i18n = get<I18n>()
             val mediaProgressFlow = MutableStateFlow<FileTransferProgress?>(null)
+            val maxPreviewSize = get<MatrixMessengerConfiguration>().filePreviewMaxSize
             launch {
                 mediaProgressFlow.collectLatest {
                     progress.emit(FileTransferProgressElement(
@@ -82,7 +87,14 @@ open class MediaViewModelImpl(
             if (encryptedFile != null) {
                 matrixClient.media.getEncryptedMedia(encryptedFile, mediaProgressFlow).fold(
                     onSuccess = {
-                        mediaDataFlow.value = it
+                        mediaDataFlow.value = it.limitSize(maxPreviewSize).catch { e ->
+                            if (e.cause is MaxByteFlowSizeException) {
+                                error.value = i18n.mediaTooLargeForPreview()
+                            } else {
+                                error.value = i18n.mediaCanNotBePreviewed()
+                            }
+                            mediaDataFlow.value = null
+                        }
                     },
                     onFailure = {
                         log.error(it) { "Cannot load encrypted ${mediaType.name} from '${encryptedFile.url}'." }