+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/files/VideoViewModel.kt
@@ -2,12 +2,21 @@ package de.connect2x.trixnity.messenger.viewmodel.files
 
 import MediaViewModel
 import MediaViewModelImpl
+import de.connect2x.trixnity.messenger.MatrixMessengerConfiguration
 import de.connect2x.trixnity.messenger.viewmodel.MatrixClientViewModelContext
+import de.connect2x.trixnity.messenger.viewmodel.i18n
 import de.connect2x.trixnity.messenger.viewmodel.room.timeline.OpenModalType
+import de.connect2x.trixnity.messenger.viewmodel.util.MaxByteFlowSizeException
+import de.connect2x.trixnity.messenger.viewmodel.util.limitSize
 import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.catch
+import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.flow.stateIn
 import net.folivo.trixnity.core.model.events.m.room.EncryptedFile
 import net.folivo.trixnity.utils.ByteArrayFlow
+import org.koin.core.component.get
 
 
 interface VideoViewModelFactory {
@@ -46,7 +55,14 @@ open class VideoViewModelImpl(
     OpenModalType.VIDEO,
     onCloseMedia,
 ), VideoViewModel {
-    override val video = mediaDataFlow
+    val maxPreviewSize = get<MatrixMessengerConfiguration>().filePreviewMaxSize
+    override val video = mediaDataFlow.map {
+        it?.limitSize(maxPreviewSize)?.catch { e ->
+            if (e.cause is MaxByteFlowSizeException) {
+                error.value = i18n.mediaTooLargeForPreview()
+            } else error.value = i18n.mediaCanNotBePreviewed()
+        }
+    }.stateIn(coroutineScope, SharingStarted.WhileSubscribed(), null)
 }
 
 class PreviewVideoViewModel : VideoViewModel {